+++ DAY04/ex02/Dog.cpp
@@ -38,7 +38,7 @@ Dog::~Dog()
 	std::cout << "D[DESCTRUCTOR] Dog is destroyed."<< std::endl;
 }
 
-std::string	Dog::get_brain_idea(int index) const {
+std::string	Dog::get_brain_idea(long int index) const {
 	return (head->get_idea(index));
 	}
